OOP SOLID - Exercise Complete

+++ C#/C# Advanced/C# OOP - June 2019/06.SOLID/Exercise/Logger/Models/Files/LogFile.cs
@@ -38,7 +38,7 @@ namespace LoggerTask.Models.Files
             ulong size = (ulong)text
                 .ToCharArray()
                 .Where(c => char.IsLetter(c))
-                .Sum(c => c);
+                .Sum(c => (int)c);
 
             return size;
         }
